v1.3.0 (January 13th, 2019) (major features)
NOW FULLY COMPATABLE WITH FG 3.3.7.

- Compatability added for FG 3.3.7 background image options.
- Layers support for background images. Updated project with new code from Enhanced Images extension, https://www.fantasygrounds.com/forums/showthread.php?20231-Enhanced-Images-(layers)-for-FG-3-0-CoreRPG-(and-rulesets-based-on-CoreRPG. [campaign/record_image.xml]
    [campaign/scripts/manager_maptoken.lua] modified code to support background image panels and to work with updated Enchaned Images code
    [campaign/record_image.xml] overwritten with code from Enhcanced Images: campaign/updated_record_image.xml
    [campaign/scripts/updated_image.lua] added, replacing older image.lua
    [campaign/scripts/updated_imagewindow.lua] added, replacing older imagewindow.lua
- Added clear saves button to all image panel types (regular/background). [scripts/manager_maptoken.lua] [campaign/record_image.xml]
- Black mask (fog of war) for all image versions, regular and background. [extension.xml, graphics/frames/imagepanel v2.png]
- Added check for left-click only activation for alt and alt+ctrl mouse clicks on tokens [modifications.lua, onClickDown]
- Changed background image buttons from Test version to smaller Live version editions by removing prior image refernces in xml, defaulting to default. [graphics/icons.xml]
- Changed image grid menu item buttons to the smaller default graphics. For better readbility and distinction from layers button graphics. [template_toolbar.xml]
- Removed a number of unneeded console outputs that happened during runtime.
- Blood splatter code changed to work with backgound image layers as well. [scripts/manager_token2.lua > function createSplatter(tokenCT,nodeCT,targetLayer), function updateStatusOverlayWidgetHelper(tokenCT,nodeCT,targetLayer)]
- Shift+left mouse click, ping and move view for players to gm focus now working with backgound images as well. [campaign/record_image.xml]
- Double-clicking Token on battlemap, now opens up character or NPC dialogue window, similarly to the default behaviour in FG. [scripts/modifications.lua > function onInit(), function onDoubleClick(tokenMap, vImage)]

+++ scripts/modifications.lua
@@ -29,7 +29,9 @@ TOKENUNDERLAYCOLOR_3 = "330000FF"; -- Token mouse over hover.
 
 
 function onInit()	
-	Token.onClickDown = onClickDown; 		
+	Token.onClickDown = onClickDown	
+	-- Token.onDrag = dynamicShadows	
+	Token.onDoubleClick = onDoubleClick;	
 end
 
 
@@ -37,9 +39,10 @@ end
 -- Use: Mouse Left-click Token on combat map while holding down either 'Alt' or 'Alt+Ctrl'
 -- Pre: Token on combat map
 -- Post: Token removed from combat map (Alt) or combat map and combat tracker (Atl+Ctrl)
+-- NOTE: button (number), Returns a numerical value indicating the button pressed (1 = left, 2 = middle, 4 = button 4, 5 = button 5). Right button is used for radial menus.
 function onClickDown( token, button, image ) 	
-	-- Deletes token from combat map, if Alt held on click.
-	if (Input.isAltPressed() == true) and (User.isHost() == true) then
+	-- Deletes token from combat map, if Alt held on left mouse click.
+	if (Input.isAltPressed() == true) and (User.isHost() == true) and (button==1) then
 		local nodeCT = CombatManager.getCTFromToken(token);
 		token.delete();		
 
@@ -50,4 +53,101 @@ function onClickDown( token, button, image )
 			end				
 		end
 	end		
+
+	-- Code below works, but for now leaving it as double-click on token, as this is the default behaviour of FG.
+	-- Opens up NPC dialogue sheet if ctrl + left mouse click on token
+	--[[
+	if (Input.isControlPressed() == true) and (User.isHost() == true) and (button==1) then		
+		--Debug.chat('open NPC sheet');
+		onDoubleClick(token, image);		
+	end	
+	]]--	
+
+	-- Allow players to move their own tokens with middle mouse button press when tokens are locked.
+	-- Possibly add this as a menu option to allow or disallow.
+	--if button == 2 then
+		-- token move on path defined during locked stage
+	--	owner = getTokenPlayer( token );
+	--end
+
+end
+
+-- a modified Token.onDrag event, for handling dynamic shadows
+function dynamicShadows( token, mouseButton, x, y, dragdata )
+	Debug.chat('token ', token)	
+	--local image = UpdatedImageWindow.image
+	Debug.chat('controls ', image)	
+	local gridType = image.getGridType() -- only work with square hexes for simplicy to begin with
+	local gridSize = image.getGridSize()
+	local gridOffset = image.getGridOffset()
+	local maskTool = image.getMaskTool()
+	local maskLayer = image.hasMask()
+
+	image.setMaskEnabled(true)
+	image.setDrawingSize(500,500, 50, 50)
+	-- make selection
+	-- mask or unmask
+	image.setMaskTool(unmaskelection) --Valid values are "maskselection" and "unmaskelection
+
+	--Debug.chat('image ', image, ' gridType ', gridType)		
+end
+
+-- Open record when token is double clicked (direct copy from CoreRPG.pak FG v3.3.7)
+function onDoubleClick(tokenMap, vImage)
+	local nodeCT = CombatManager.getCTFromToken(tokenMap);
+	if nodeCT then
+		if User.isHost() then
+			local sClass, sRecord = DB.getValue(nodeCT, "link", "", "");
+			if sRecord ~= "" then
+				Interface.openWindow(sClass, sRecord);				
+			else
+				Interface.openWindow(sClass, nodeCT);				
+			end
+		else
+			if (DB.getValue(nodeCT, "friendfoe", "") == "friend") then
+				local sClass, sRecord = DB.getValue(nodeCT, "link", "", "");
+				if sClass == "charsheet" then
+					if sRecord ~= "" and DB.isOwner(sRecord) then
+						Interface.openWindow(sClass, sRecord);						
+					else
+						ChatManager.SystemMessage(Interface.getString("ct_error_openpclinkedtokenwithoutaccess"));
+					end
+				else
+					local nodeActor;
+					if sRecord ~= "" then
+						nodeActor = DB.findNode(sRecord);
+					else
+						nodeActor = nodeCT;
+					end
+					if nodeActor then
+						Interface.openWindow(sClass, nodeActor);						
+					else
+						ChatManager.SystemMessage(Interface.getString("ct_error_openotherlinkedtokenwithoutaccess"));
+					end
+				end
+				vImage.clearSelectedTokens();
+			end
+		end
+	end
+end
+
+-- Ctrl click on map, timed function that draws circles ever larger (up to a set point), while deleting previous, to indicate a pinged location. 
+-- Giving the impression of a timed expanding circle.
+function pingExpandingCircle()
+	-- PointerToolkit  using scripts/pointer_toolkit.lua
+end
+
+-- Return owner for character sheet connected to Token
+function getTokenPlayer(token)	
+	local nodeCT = CombatManager.getCTFromToken(token);
+	if nodeCT then
+		local owner = nodeCT.getOwner();
+		Debug.console('modifications>getTokenPlayer, nodeCt owner:', owner);
+		--- DB.isOwner( nodeid )
+		
+		if User.isHost() then
+		end	
+		if not User.isHost() then
+		end
+	end				
 end
